Update stripex tool to support new PRG+ELF program format

+++ sys/usb/src.km/tos-common/stripex.c
@@ -1,6 +1,6 @@
 /*
  * stripex.c 
- * strip symbol table and GNU binutils aexec header from TOS executables.
+ * strip symbol table, GNU binutils aexec header, and ELF headers from TOS executables.
  * Needed e.g. for SLB shared library files or CPX modules to get the header back
  * to the start of the text segment.
  *
@@ -44,17 +44,31 @@ struct aexec
 	unsigned long a_data;				/* size of initialized data */
 	unsigned long a_bss;				/* size of uninitialized data */
 	unsigned long a_syms;				/* size of symbol table */
-	unsigned long a_AZero1;				/* always zero */
+	unsigned long a_res1;				/* see below */
 	unsigned long a_AZero2;				/* always zero */
 	unsigned short a_isreloc;			/* is reloc info present */
 };
 
 #define	CMAGIC	0x601A					/* contiguous text */
 
+/*
+ * magic values in a_res1 field
+ */
+#define EXEC_FMT_TOS  0 /* original TOS */
+#define EXEC_FMT_MINT 0x4d694e54UL /* binutils a.out format */
+#define EXEC_FMT_ELF  0x454c4600UL /* new binutils PRG+ELF format */
+
+
 #define SIZEOF_SHORT 2
 #define SIZEOF_LONG  4
 
+/*
+ * sizeof(struct aexec)
+ */
 #define SIZEOF_AEXEC (SIZEOF_SHORT + (6 * SIZEOF_LONG) + SIZEOF_SHORT)
+/*
+ * extra header added by a.out-mintprg format
+ */
 #define SIZEOF_BINUTILS_AEXEC 228
 
 
@@ -116,7 +130,7 @@ static int read_head(int fd, struct aexec *a)
 	a->a_data = read_belong(buf + 6);
 	a->a_bss = read_belong(buf + 10);
 	a->a_syms = read_belong(buf + 14);
-	a->a_AZero1 = read_belong(buf + 18);
+	a->a_res1 = read_belong(buf + 18);
 	a->a_AZero2 = read_belong(buf + 22);
 	a->a_isreloc = read_beword(buf + 26);
 
@@ -133,7 +147,7 @@ static int write_head(int fd, const struct aexec *a)
 	write_belong(buf + 6, a->a_data);
 	write_belong(buf + 10, a->a_bss);
 	write_belong(buf + 14, a->a_syms);
-	write_belong(buf + 18, a->a_AZero1);
+	write_belong(buf + 18, a->a_res1);
 	write_belong(buf + 22, a->a_AZero2);
 	write_beword(buf + 26, a->a_isreloc);
 
@@ -149,8 +163,8 @@ static int write_head(int fd, const struct aexec *a)
  */
 static long copy(int from, int to, long bytes)
 {
-	register long todo;
-	long done = 0L;
+	long todo;
+	long done = 0;
 	long remaining = bytes;
 	long actual;
 
@@ -179,7 +193,7 @@ static long copy(int from, int to, long bytes)
 }
 
 
-static int relocate(const char *name, int fd, long reloc_pos)
+static int relocate(const char *name, int fd, long reloc_pos, long bytes_to_delete)
 {
 	if (lseek(fd, reloc_pos, SEEK_SET) != reloc_pos)
 	{
@@ -191,7 +205,7 @@ static int relocate(const char *name, int fd, long reloc_pos)
 		perror(name);
 		return 0;
 	}
-	write_belong(mybuf, read_belong(mybuf) - SIZEOF_BINUTILS_AEXEC);
+	write_belong(mybuf, read_belong(mybuf) - bytes_to_delete);
 	if (lseek(fd, reloc_pos, SEEK_SET) != reloc_pos)
 	{
 		perror(name);
@@ -207,16 +221,17 @@ static int relocate(const char *name, int fd, long reloc_pos)
 
 
 /* copy TOS relocation table from `from` to `to`. Copy bytes until NUL byte or
-   first 4 bytes if == 0l.
+   first 4 bytes if == 0.
    returns length of relocation table or -1 in case of an error */
 
-static long copy_relocs(const char *name, int from, int to)
+static long copy_relocs(const char *name, int from, int to, long bytes_to_delete, long image_size)
 {
 	long res = 0;
 	long bytes;
 	long rbytes = 0;
 	long first_relo;
 	long reloc_pos;
+	long limit;
 	
 	res = read(from, mybuf, SIZEOF_LONG);
 	if (res != 0 && res != SIZEOF_LONG)
@@ -235,7 +250,20 @@ static long copy_relocs(const char *name, int from, int to)
 	first_relo = read_belong(mybuf);
 	if (first_relo != 0)
 	{
-		first_relo -= SIZEOF_BINUTILS_AEXEC;
+		/*
+		 * The initially designed elf format had some relocations in the header
+		 * that we are about to remove.
+		 * Supporting that would require to rebuild the relocation table.
+		 * The current format does not have such relocations anymore,
+		 * but better check for it.
+		 */
+		if (first_relo <= bytes_to_delete)
+		{
+			fprintf(stderr, "%s: first relocation at 0x%08lx is in area to be deleted\n", name, first_relo);
+			fprintf(stderr, "update your binutils\n");
+			return -1;
+		}
+		first_relo -= bytes_to_delete;
 		write_belong(mybuf, first_relo);
 	}
 	
@@ -251,12 +279,13 @@ static long copy_relocs(const char *name, int from, int to)
 										   relocation table                   */
 
 	reloc_pos = SIZEOF_AEXEC + first_relo;
-	if (relocate(name, to, reloc_pos) == 0)
+	if (relocate(name, to, reloc_pos, bytes_to_delete) == 0)
 		return -1;
+	limit = SIZEOF_AEXEC + image_size;
 
 	for (;;)
 	{
-		lseek(to, 0l, SEEK_END);
+		lseek(to, 0, SEEK_END);
 		if ((bytes = read(from, mybuf, 1)) < 0)
 		{
 			fprintf(stderr, "%s: Error reading\n", name);
@@ -281,7 +310,17 @@ static long copy_relocs(const char *name, int from, int to)
 		} else
 		{
 			reloc_pos += mybuf[0];
-			if (relocate(name, to, reloc_pos) == 0)
+			/*
+			 * check that we don't try to write beyond the new image size.
+			 * This too, could happen with early version of new elf binutils
+			 */
+			if (reloc_pos >= limit)
+			{
+				fprintf(stderr, "%s: relocation at 0x%08lx is outside image\n", name, reloc_pos);
+				fprintf(stderr, "update your binutils\n");
+				return -1;
+			}
+			if (relocate(name, to, reloc_pos, bytes_to_delete) == 0)
 				return -1;
 		}
 	}
@@ -291,13 +330,14 @@ static long copy_relocs(const char *name, int from, int to)
 
 static int strip(const char *name)
 {
-	register int fd;
-	register int tfd;
-	register long count, rbytes, sbytes;
+	int fd;
+	int tfd;
+	long count, rbytes, sbytes;
 	struct aexec ahead;
-	unsigned char buf[8 * SIZEOF_LONG];
-	long magic1, magic2;
-	
+	unsigned char buf[13 * SIZEOF_LONG];
+	unsigned long magic1, magic2;
+	long bytes_to_delete;
+
 	if ((fd = open(name, O_RDONLY | O_BINARY, 0755)) < 0)
 	{
 		perror(name);
@@ -311,10 +351,10 @@ static int strip(const char *name)
 	}
 	
 	/*
-	 * read g_jump_entry and first 6 longs of exec header
+	 * read g_jump_entry and first 8 longs of exec header
 	 */
 	if (read_head(fd, &ahead) < 0 ||
-		(int)read(fd, buf, sizeof(buf)) != (int)sizeof(buf))
+		(int)read(fd, buf, 8 * SIZEOF_LONG) != 8 * SIZEOF_LONG)
 	{
 		perror(name);
 		close(tfd);
@@ -328,43 +368,148 @@ static int strip(const char *name)
 		close(fd);
 		return 1;
 	}
-	magic1 = read_belong(buf);
-	magic2 = read_belong(buf + SIZEOF_LONG);
-	if (!((magic1 == 0x283a001aL && magic2 == 0x4efb48faL) ||	/* Original binutils */
-		  (magic1 == 0x203a001aL && magic2 == 0x4efb08faL)))		/* binutils >= 2.18-mint-20080209 */
+	if (ahead.a_res1 == EXEC_FMT_MINT)		/* 'MiNT' extended exec header magic */
 	{
-		fprintf(stderr, "%s: no aexec header\n", name);
-		close(tfd);
-		close(fd);
-		return 1;
-	}
-#if 0
-	magic1 = read_belong(buf + 2 * SIZEOF_LONG);
-	printf("e_info: %08lx\n", magic1);
-	magic1 = read_belong(buf + 3 * SIZEOF_LONG);
-	printf("e_text: %08lx\n", magic1);
-	magic1 = read_belong(buf + 4 * SIZEOF_LONG);
-	printf("e_data: %08lx\n", magic1);
-	magic1 = read_belong(buf + 5 * SIZEOF_LONG);
-	printf("e_bss: %08lx\n", magic1);
-	magic1 = read_belong(buf + 6 * SIZEOF_LONG);
-	printf("e_syms: %08lx\n", magic1);
-#endif
-	magic1 = read_belong(buf + 7 * SIZEOF_LONG);
-#if 0
-	printf("e_entry: %08lx\n", magic1);
-#endif
-	if (magic1 != SIZEOF_BINUTILS_AEXEC)
-	{
-		fprintf(stderr, "%s: warning: entry %08lx not at start of text segment\n", name, magic1);
-		if (!force)
+		long e_entry;
+
+		bytes_to_delete = SIZEOF_BINUTILS_AEXEC;
+
+		/*
+		 * read trampoline code
+		 */
+		magic1 = read_belong(buf);
+		magic2 = read_belong(buf + SIZEOF_LONG);
+		if (!((magic1 == 0x283a001aUL && magic2 == 0x4efb48faUL) ||	/* Original binutils */
+			  (magic1 == 0x203a001aUL && magic2 == 0x4efb08faUL)))		/* binutils >= 2.18-mint-20080209 */
 		{
-			fprintf(stderr, "use -f to override\n");
+			fprintf(stderr, "%s: no aexec header\n", name);
 			close(tfd);
 			close(fd);
 			return 1;
 		}
+#if 0
+		magic1 = read_belong(buf + 2 * SIZEOF_LONG);
+		printf("e_info: %08lx\n", magic1);
+		magic1 = read_belong(buf + 3 * SIZEOF_LONG);
+		printf("e_text: %08lx\n", magic1);
+		magic1 = read_belong(buf + 4 * SIZEOF_LONG);
+		printf("e_data: %08lx\n", magic1);
+		magic1 = read_belong(buf + 5 * SIZEOF_LONG);
+		printf("e_bss: %08lx\n", magic1);
+		magic1 = read_belong(buf + 6 * SIZEOF_LONG);
+		printf("e_syms: %08lx\n", magic1);
+#endif
+		e_entry = read_belong(buf + 7 * SIZEOF_LONG);
+#if 0
+		printf("e_entry: %08lx\n", e_entry);
+#endif
+		if (e_entry != bytes_to_delete)
+		{
+			fprintf(stderr, "%s: warning: entry %08lx not at start of text segment\n", name, e_entry);
+			if (!force)
+			{
+				fprintf(stderr, "use -f to override\n");
+				close(tfd);
+				close(fd);
+				return 1;
+			}
+		}
+	} else if ((ahead.a_res1 & 0xffffff00UL) == EXEC_FMT_ELF && (ahead.a_res1 & 0xff) >= 40)
+	{
+		long elf_offset;
+		long e_phoff;
+		long e_entry;
+		long p_offset;
+		unsigned short e_phnum;
+		
+		/*
+		 * offset of ELF header is determined by last byte of a_res1 (normally 40)
+		 */
+		elf_offset = (ahead.a_res1 & 0xff);
+		/*
+		 * read trampoline code
+		 */
+		magic1 = read_belong(buf);
+		magic2 = read_belong(buf + SIZEOF_LONG);
+		if (magic1 != (0x203a0000UL | (elf_offset + 24 - 30)) || magic2 != 0x4efb08faUL)		/* binutils >= 2.41-mintelf */
+		{
+			fprintf(stderr, "%s: no exec header\n", name);
+			close(tfd);
+			close(fd);
+			return 1;
+		}
+#if 0
+		magic1 = read_belong(buf + 2 * SIZEOF_LONG);
+		printf("_stksize: @%08lx\n", magic1);
+#endif
+		if (lseek(fd, elf_offset, SEEK_SET) < 0 ||
+			(int)read(fd, buf, 13 * SIZEOF_LONG) != 13 * SIZEOF_LONG)
+		{
+			perror(name);
+			close(tfd);
+			close(fd);
+			return 1;
+		}
+		/*
+		 * We will find a complete ELF header there,
+		 * but only care here about e_magic & e_entry
+		 */
+		magic1 = read_belong(buf);
+		e_phnum = read_beword(buf + 44);
+		if (magic1 != 0x7f454c46UL || e_phnum == 0)
+		{
+			fprintf(stderr, "%s: no ELF header\n", name);
+			close(tfd);
+			close(fd);
+			return 1;
+		}
+		e_entry = read_belong(buf + 6 * SIZEOF_LONG);
+		e_phoff = read_belong(buf + 7 * SIZEOF_LONG);
+		/*
+		 * read first program header
+		 */
+		if (lseek(fd, e_phoff, SEEK_SET) < 0 ||
+			(int)read(fd, buf, 8 * SIZEOF_LONG) != 8 * SIZEOF_LONG)
+		{
+			perror(name);
+			close(tfd);
+			close(fd);
+			return 1;
+		}
+		/*
+		 * get start of text segment
+		 */
+		p_offset = read_belong(buf + SIZEOF_LONG);
+		bytes_to_delete = e_phoff + e_phnum * 32 - SIZEOF_AEXEC;
+		if (p_offset != SIZEOF_AEXEC || e_entry != bytes_to_delete)
+		{
+			fprintf(stderr, "%s: warning: entry %08lx not at start of text segment %08lx\n", name, e_entry, bytes_to_delete);
+			if (!force)
+			{
+				fprintf(stderr, "use -f to override\n");
+				close(tfd);
+				close(fd);
+				return 1;
+			}
+		}
+	} else if (ahead.a_res1 == EXEC_FMT_TOS)
+	{
+		if (ahead.a_syms == 0)
+		{
+			fprintf(stderr, "%s: already stripped\n", name);
+			close(tfd);
+			close(fd);
+			return 0;
+		}
+		bytes_to_delete = 0;
+	} else
+	{
+		fprintf(stderr, "%s: unsupported file format\n", name);
+		close(tfd);
+		close(fd);
+		return 1;
 	}
+
 	sbytes = ahead.a_syms;
 	if (verbose)
 	{
@@ -372,25 +517,25 @@ static int strip(const char *name)
 	}
 
 	ahead.a_syms = 0;
-	ahead.a_text -= SIZEOF_BINUTILS_AEXEC;
-	if (ahead.a_AZero1 == 0x4d694e54l)		/* 'MiNT' extended exec header magic */
-		ahead.a_AZero1 = 0;
+	ahead.a_text -= bytes_to_delete;
+	ahead.a_res1 = 0;
 	if (write_head(tfd, &ahead) < 0)
 	{
+		perror(name);
 		close(tfd);
 		close(fd);
 		return 1;
 	}
-	if (lseek(fd, SIZEOF_BINUTILS_AEXEC - sizeof(buf), SEEK_CUR) < 0)
+	if (lseek(fd, bytes_to_delete + SIZEOF_AEXEC, SEEK_SET) < 0)
 	{
-		fprintf(stderr, "%s: seek error\n", name);
+		perror(name);
 		close(tfd);
 		close(fd);
 		return 1;
 	}
 	if (verbose)
 	{
-		printf("%s: skipped 0x%x bytes aexec header\n", name, SIZEOF_BINUTILS_AEXEC);
+		printf("%s: skipped 0x%lx bytes extra header\n", name, bytes_to_delete);
 	}
 
 	count = ahead.a_text + ahead.a_data;
@@ -406,7 +551,7 @@ static int strip(const char *name)
 	}
 	if (lseek(fd, sbytes, SEEK_CUR) < 0)
 	{
-		fprintf(stderr, "%s: seek error\n", name);
+		perror(name);
 		close(tfd);
 		close(fd);
 		return 1;
@@ -418,7 +563,7 @@ static int strip(const char *name)
 	rbytes = 0;
 	if (ahead.a_isreloc == 0)
 	{
-		if ((rbytes = copy_relocs(name, fd, tfd)) < 0)
+		if ((rbytes = copy_relocs(name, fd, tfd, bytes_to_delete, count)) < 0)
 		{
 			close(tfd);
 			close(fd);
@@ -433,12 +578,13 @@ static int strip(const char *name)
 	{
 		unsigned long pos, size;
 		
-		pos = lseek(fd, 0L, SEEK_CUR);
-		lseek(fd, 0L, SEEK_END);
-		size = lseek(fd, 0L, SEEK_CUR);
+		pos = lseek(fd, 0, SEEK_CUR);
+		lseek(fd, 0, SEEK_END);
+		size = lseek(fd, 0, SEEK_CUR);
 		lseek(fd, pos, SEEK_SET);
 		if (size > pos)
 		{
+			/* could be pure-c debug information, after end of relocation table */
 			printf("%s: skipped 0x%lx bytes trailer\n", name, size - pos);
 		}
 	}
@@ -476,8 +622,7 @@ static void usage(const char *s)
 {
 	fprintf(stderr, "%s", s);
 	fprintf(stderr, "Usage: stripex [-f] files ...\n");
-	fprintf(stderr, "strip GNU-binutils aexec header from executables\n");
-	exit(1);
+	fprintf(stderr, "strip GNU-binutils aexec/elf header from executables\n");
 }
 
 
@@ -503,26 +648,36 @@ int main(int argc, char **argv)
 			break;
 		default:
 			usage("");
-			break;
+			return 1;
 		}
 	}
 
 	if (argc < 1)
 	{
 		usage("");
+		return 1;
 	}
 	
 	do
 	{
 		const char *filename = *argv++;
 		char *base;
-		
+		int fd;
+
 		strcpy(tmpname, filename);
 		base = f_basename(tmpname);
 		strcpy(base, "STXXXXXX");
-		mkstemp(tmpname);
-		status |= strip(filename);
-		unlink(tmpname);
+		fd = mkstemp(tmpname);
+		if (fd < 0)
+		{
+			perror(tmpname);
+			status |= 1;
+		} else
+		{
+			close(fd);
+			status |= strip(filename);
+			unlink(tmpname);
+		}
 	} while (--argc > 0);
 
 	return status;
